Show dependencies of selected commits
When one or more commits are selected, highlight other commits
that touch the same files.
This helps the user to know how far he can move a commit up/down
without causing a conflict.

It also simplifies applying a fixup to a older commit
further down in the list, because it is easier to find.

Signed-off-by: Peter Oberndorfer <kumbayo84@arcor.de>